requirements.txt
Flask==3.0.3
gspread
google-auth
gunicorn
requests
